Remove crash as a direction and special case it.
Prevent setting a direction past the extremes of left or right.

The bug was being caused because turning left from the crash "direction" was an invalid direction. Instead of just putting a block in place to fix the bug, it made more sense to treat crash as a separate state rather than just another direction.