Add jump ramps and jumping
* Reference jump asset constants
* Start, stop, and animate a jump
* Detect collisions based on obstacle type and jumping state
* Add unit tests for collision detection
* Slowed the skier down a bit for better control

After moving around the same area, so many obstacles were placed that there was no more room. Because of how open positions were calculated recursively, this could lead to a stack overflow. The solution was to limit retries for an open position to 10 times before giving up and not placing another obstacle.